+++ navajo-cli/src/encoding.rs
@@ -1,8 +1,9 @@
 use std::collections::VecDeque;
 use std::io::{Read, Write};
+use std::mem;
 use std::str::FromStr;
 
-use anyhow::{bail, Context};
+use anyhow::{bail, Context, Result};
 use base64::engine::{
     general_purpose::{STANDARD as BASE64_STANDARD, URL_SAFE as BASE64_URL_SAFE},
     DecodePaddingMode, GeneralPurpose, GeneralPurposeConfig,
@@ -53,7 +54,7 @@ impl Encoding {
         }
     }
 
-    pub fn decode<T>(&self, value: T) -> anyhow::Result<Vec<u8>>
+    pub fn decode<T>(&self, value: T) -> Result<Vec<u8>>
     where
         T: AsRef<[u8]>,
     {
@@ -65,22 +66,23 @@ impl Encoding {
     }
     pub fn encode_writer<W: Write>(&self, write: W, pad: bool) -> EncodingWriter<W> {
         match self {
-            Encoding::Hex => EncodingWriter::Hex(HexWriter::new(write)),
+            Encoding::Hex => EncodingWriter::Hex(HexWriter::new(write).into()),
             Encoding::Base64 => {
                 if pad {
-                    EncodingWriter::Base64(Base64Writer::new(write, &BASE64_STANDARD))
+                    EncodingWriter::Base64(Base64Writer::new(write, &BASE64_STANDARD).into())
                 } else {
-                    EncodingWriter::Base64(Base64Writer::new(write, &BASE64_STANDARD_NO_PAD_INDIF))
+                    EncodingWriter::Base64(
+                        Base64Writer::new(write, &BASE64_STANDARD_NO_PAD_INDIF).into(),
+                    )
                 }
             }
             Encoding::Base64url => {
                 if pad {
-                    EncodingWriter::Base64Url(Base64Writer::new(write, &BASE64_URL_SAFE))
+                    EncodingWriter::Base64Url(Base64Writer::new(write, &BASE64_URL_SAFE).into())
                 } else {
-                    EncodingWriter::Base64Url(Base64Writer::new(
-                        write,
-                        &BASE64_URL_SAFE_NO_PAD_INDIF,
-                    ))
+                    EncodingWriter::Base64Url(
+                        Base64Writer::new(write, &BASE64_URL_SAFE_NO_PAD_INDIF).into(),
+                    )
                 }
             }
         }
@@ -114,15 +116,15 @@ impl Encoding {
     fn encode_hex(value: &[u8]) -> String {
         hex::encode(value)
     }
-    fn decode_base64(value: &[u8]) -> anyhow::Result<Vec<u8>> {
+    fn decode_base64(value: &[u8]) -> Result<Vec<u8>> {
         BASE64_STANDARD_NO_PAD_INDIF
             .decode(value)
             .context("failed to decode base64")
     }
-    fn decode_hex(value: &[u8]) -> anyhow::Result<Vec<u8>> {
+    fn decode_hex(value: &[u8]) -> Result<Vec<u8>> {
         hex::decode(value).context("failed to decode hex")
     }
-    fn decode_base64_url(value: &[u8]) -> anyhow::Result<Vec<u8>> {
+    fn decode_base64_url(value: &[u8]) -> Result<Vec<u8>> {
         BASE64_URL_SAFE_NO_PAD_INDIF
             .decode(value)
             .context("failed to decode base64")
@@ -136,6 +138,7 @@ where
     Base64(Base64Reader<'static, GeneralPurpose, R>),
     Base64url(Base64Reader<'static, GeneralPurpose, R>),
     Hex(HexReader<R>),
+    None(R),
 }
 impl<R: Read> Read for EncodingReader<R> {
     fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
@@ -143,6 +146,7 @@ impl<R: Read> Read for EncodingReader<R> {
             EncodingReader::Base64(r) => r.read(buf),
             EncodingReader::Base64url(r) => r.read(buf),
             EncodingReader::Hex(r) => r.read(buf),
+            EncodingReader::None(r) => r.read(buf),
         }
     }
 }
@@ -158,29 +162,98 @@ pub enum EncodingWriter<W>
 where
     W: Write,
 {
-    Base64(Base64Writer<'static, GeneralPurpose, W>),
-    Base64Url(Base64Writer<'static, GeneralPurpose, W>),
-    Base64Padded(Base64Writer<'static, GeneralPurpose, W>),
-    Base64UrlPadded(Base64Writer<'static, GeneralPurpose, W>),
-    Hex(HexWriter<W>),
+    Base64(Option<Base64Writer<'static, GeneralPurpose, W>>),
+    Base64Url(Option<Base64Writer<'static, GeneralPurpose, W>>),
+    Base64Padded(Option<Base64Writer<'static, GeneralPurpose, W>>),
+    Base64UrlPadded(Option<Base64Writer<'static, GeneralPurpose, W>>),
+    Hex(Option<HexWriter<W>>),
+    None(Option<W>),
+}
+
+impl<W: Write> EncodingWriter<W> {
+    pub fn into_inner(mut self) -> Result<W> {
+        match self {
+            EncodingWriter::Base64(ref mut w) => {
+                let mut w = w.take().context("writer already closed")?;
+                w.flush()?;
+                w.finish()?;
+                Ok(w.into_inner())
+            }
+            EncodingWriter::Base64Url(ref mut w) => {
+                let mut w = w.take().context("writer already closed")?;
+                w.flush()?;
+                w.finish()?;
+                Ok(w.into_inner())
+            }
+            EncodingWriter::Hex(ref mut w) => {
+                let mut w = w.take().context("writer already closed")?;
+                w.flush()?;
+                w.into_inner()
+            }
+            EncodingWriter::Base64Padded(ref mut w) => {
+                let mut w = w.take().context("writer already closed")?;
+                w.flush()?;
+                w.finish()?;
+                Ok(w.into_inner())
+            }
+            EncodingWriter::Base64UrlPadded(ref mut w) => {
+                let mut w = w.take().context("writer already closed")?;
+                w.flush()?;
+                w.finish()?;
+                Ok(w.into_inner())
+            }
+            EncodingWriter::None(ref mut w) => {
+                let w = w.take().context("writer already closed")?;
+                Ok(w)
+            }
+        }
+    }
 }
 impl<W: Write> Write for EncodingWriter<W> {
     fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
         match self {
-            EncodingWriter::Base64(w) => w.write(buf),
-            EncodingWriter::Base64Url(w) => w.write(buf),
-            EncodingWriter::Hex(w) => w.write(buf),
-            EncodingWriter::Base64Padded(w) => w.write(buf),
-            EncodingWriter::Base64UrlPadded(w) => w.write(buf),
+            EncodingWriter::Base64(ref mut w) => {
+                check_if_closed(w)?;
+                w.as_mut().unwrap().write(buf)
+            }
+            EncodingWriter::Base64Url(ref mut w) => {
+                check_if_closed(w)?;
+                w.as_mut().unwrap().write(buf)
+            }
+            EncodingWriter::Hex(ref mut w) => {
+                check_if_closed(w)?;
+                w.as_mut().unwrap().write(buf)
+            }
+            EncodingWriter::Base64Padded(ref mut w) => {
+                check_if_closed(w)?;
+                w.as_mut().unwrap().write(buf)
+            }
+            EncodingWriter::Base64UrlPadded(ref mut w) => {
+                check_if_closed(w)?;
+                w.as_mut().unwrap().write(buf)
+            }
+            EncodingWriter::None(ref mut w) => {
+                check_if_closed(w)?;
+                w.as_mut().unwrap().write(buf)
+            }
         }
     }
     fn flush(&mut self) -> std::io::Result<()> {
         match self {
-            EncodingWriter::Base64(w) => w.flush(),
-            EncodingWriter::Base64Url(w) => w.flush(),
-            EncodingWriter::Hex(w) => w.flush(),
-            EncodingWriter::Base64Padded(w) => w.flush(),
-            EncodingWriter::Base64UrlPadded(w) => w.flush(),
+            EncodingWriter::Base64(w) => {
+                w.as_mut().and_then(|w| w.flush().into()).unwrap_or(Ok(()))
+            }
+            EncodingWriter::Base64Url(w) => {
+                w.as_mut().and_then(|w| w.flush().into()).unwrap_or(Ok(()))
+            }
+            EncodingWriter::Hex(w) => w.as_mut().and_then(|w| w.flush().into()).unwrap_or(Ok(())),
+            EncodingWriter::Base64Padded(w) => {
+                w.as_mut().and_then(|w| w.flush().into()).unwrap_or(Ok(()))
+            }
+            EncodingWriter::Base64UrlPadded(w) => {
+                w.as_mut().and_then(|w| w.flush().into()).unwrap_or(Ok(()))
+            }
+            EncodingWriter::None(w) => w.as_mut().and_then(|w| w.flush().into()).unwrap_or(Ok(())),
         }
     }
 }
@@ -190,26 +263,41 @@ impl<W: Write> Drop for EncodingWriter<W> {
     fn drop(&mut self) {
         match self {
             EncodingWriter::Base64(w) => {
-                if w.flush().is_ok() {
-                    w.finish();
+                if let Some(w) = w {
+                    if w.flush().is_ok() {
+                        w.finish();
+                    }
                 }
             }
             EncodingWriter::Base64Url(w) => {
-                if w.flush().is_ok() {
-                    w.finish();
+                if let Some(w) = w {
+                    if w.flush().is_ok() {
+                        w.finish();
+                    }
                 }
             }
             EncodingWriter::Hex(w) => {
-                w.flush();
+                if let Some(w) = w {
+                    w.flush();
+                }
             }
             EncodingWriter::Base64Padded(w) => {
-                if w.flush().is_ok() {
-                    w.finish();
+                if let Some(w) = w {
+                    if w.flush().is_ok() {
+                        w.finish();
+                    }
                 }
             }
             EncodingWriter::Base64UrlPadded(w) => {
-                if w.flush().is_ok() {
-                    w.finish();
+                if let Some(w) = w {
+                    if w.flush().is_ok() {
+                        w.finish();
+                    }
+                }
+            }
+            EncodingWriter::None(w) => {
+                if let Some(w) = w {
+                    w.flush();
                 }
             }
         };
@@ -231,41 +319,102 @@ impl<R: Read> Read for HexReader<R> {
     }
 }
 pub struct HexWriter<W: Write> {
-    w: W,
+    w: Option<W>,
     buf: VecDeque<u8>,
+    remnant: Option<u8>,
 }
 impl<W: Write> HexWriter<W> {
     pub fn new(w: W) -> Self {
         Self {
-            w,
+            w: Some(w),
             buf: VecDeque::new(),
+            remnant: None,
         }
+    }
+    /// Flushes the internal buffer and returns the underlying writer.
+    fn into_inner(mut self) -> Result<W, anyhow::Error> {
+        self.flush()?;
+        self.w
+            .take()
+            .ok_or(anyhow::anyhow!("writer is already closed"))
     }
 }
 impl<W: Write> Write for HexWriter<W> {
     fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        self.buf.reserve(buf.len() * 2);
-        for b in hex::encode(buf).as_bytes() {
-            self.buf.push_back(*b);
+        check_if_closed(&self.w)?;
+
+        let mut remaining = buf.len();
+        let mut encoded = Vec::with_capacity(remaining);
+
+        if let Some(remnant) = self.remnant.take() {
+            encoded.push(remnant);
+            remaining -= 1;
         }
-        let mut i = 0;
-        let mut data = vec![0; self.buf.len()];
-        while i < buf.len() && !self.buf.is_empty() {
-            data[i] = self.buf.pop_front().unwrap();
-            i += 1;
+        self.buf.extend(buf.iter());
+        while remaining > 0 && !self.buf.is_empty() {
+            if remaining > 1 {
+                encoded.extend([0, 0].iter());
+                let len = encoded.len();
+                hex::encode_to_slice([self.buf.pop_front().unwrap()], &mut encoded[len - 2..])
+                    .map_err(|e| {
+                        std::io::Error::new(
+                            std::io::ErrorKind::InvalidData,
+                            format!("failed to encode hex: {}", e),
+                        )
+                    })?;
+                remaining -= 2;
+            } else {
+                let mut last_bytes = [0u8, 0u8];
+                hex::encode_to_slice([self.buf.pop_front().unwrap()], &mut last_bytes).map_err(
+                    |e| {
+                        std::io::Error::new(
+                            std::io::ErrorKind::InvalidData,
+                            format!("failed to encode hex: {}", e),
+                        )
+                    },
+                )?;
+                encoded.push(last_bytes[0]);
+                self.remnant = Some(last_bytes[1]);
+                break;
+            }
         }
-        self.w.write(&data[0..i])
+        self.w.as_mut().unwrap().write(&encoded)
     }
 
     fn flush(&mut self) -> std::io::Result<()> {
-        self.w.flush()
+        if self.w.is_none() {
+            return Err(std::io::Error::new(
+                std::io::ErrorKind::Other,
+                "writer is already closed",
+            ));
+        }
+        let mut remaining = vec![0u8; self.buf.len() * 2];
+        let mut start = 0;
+        if let Some(remnant) = self.remnant.take() {
+            remaining.push(0);
+            remaining[0] = remnant;
+            start = 1;
+        }
+        let mut buf = mem::take(&mut self.buf);
+        hex::encode_to_slice(buf.make_contiguous(), &mut remaining[start..]).map_err(|e| {
+            std::io::Error::new(
+                std::io::ErrorKind::InvalidData,
+                format!("failed to encode hex: {}", e),
+            )
+        })?;
+        let w = self.w.as_mut().unwrap();
+        w.write_all(&remaining)?;
+        w.flush()
     }
 }
 
 #[allow(unused_must_use)]
 impl<W: Write> Drop for HexWriter<W> {
     fn drop(&mut self) {
-        self.w.write(self.buf.make_contiguous()); // error is ignored
+        let w = self.w.take();
+        if let Some(mut w) = w {
+            w.write_all(self.buf.make_contiguous()); // error is ignored
+        }
     }
 }
 
@@ -283,6 +432,17 @@ impl FromStr for Encoding {
             "base64url" => Ok(Self::Base64url),
             _ => bail!("unknown encoding: {}", s),
         }
+    }
+}
+
+fn check_if_closed<T>(w: &Option<T>) -> Result<(), std::io::Error> {
+    if w.is_none() {
+        Err(std::io::Error::new(
+            std::io::ErrorKind::Other,
+            "writer already closed",
+        ))
+    } else {
+        Ok(())
     }
 }
 
@@ -429,7 +589,7 @@ mod tests {
         let rng = navajo::rand::SystemRng::default();
         let mut count = 0;
         while count == 0 {
-            count = rng.u8().unwrap();
+            count = rng.u8().unwrap() % 10;
         }
         for _ in 0..count {
             let mut size = 0;